Testing random colors for lines

+++ app/src/main/java/good/damn/marqueeview/utils/FileUtils.java
@@ -43,7 +43,7 @@ public class FileUtils {
                 fos.write(ByteUtils.fixedPointNumber(l.toX));
                 fos.write(ByteUtils.fixedPointNumber(l.toY));
                 fos.write(ByteUtils.integer(l.color));
-                fos.write((byte) l.strokeWidth);
+                fos.write(l.strokeWidth);
             }
 
             fos.close();
